9818: DRY up internal case parties helper functions
