+++ web-client/src/presenter/computeds/internalPetitionPartiesHelper.ts
@@ -195,19 +195,26 @@ export const getOptionsForContact = ({ PARTY_TYPES, partyType }) => {
 };
 
 /**
- * gets the contact view options based on partyType
+ * gets the contact view options based on form.partyType
  *
  * @param {Function} get the cerebral get function used
- * @param {object} applicationContext the application context
  * for getting state.caseDetail.partyType and state.constants
+ * @param {object} applicationContext the application context
  * @returns {object} the contactPrimary and/or contactSecondary
  * view options
  */
-export const caseDetailEditContactsHelper = (get, applicationContext) => {
+export const internalPetitionPartiesHelper = (get, applicationContext) => {
   const partyType = get(state.form.partyType);
+  const user = applicationContext.getCurrentUser();
+
   const { PARTY_TYPES } = applicationContext.getConstants();
+  const isExternalUser = applicationContext
+    .getUtilities()
+    .isExternalUser(user.role);
+
+  let showPaperPetitionEmailFieldAndConsentBox = !isExternalUser;
 
   const contacts = getOptionsForContact({ PARTY_TYPES, partyType });
 
-  return contacts;
+  return { ...contacts, showPaperPetitionEmailFieldAndConsentBox };
 };
