--- web-client/src/presenter/computeds/startCaseInternalContactsHelper.ts
@@ -1,27 +1,0 @@
-import { getOptionsForContact } from './caseDetailEditContactsHelper';
-import { state } from 'cerebral';
-
-/**
- * gets the contact view options based on form.partyType
- *
- * @param {Function} get the cerebral get function used
- * for getting state.caseDetail.partyType and state.constants
- * @param {object} applicationContext the application context
- * @returns {object} the contactPrimary and/or contactSecondary
- * view options
- */
-export const startCaseInternalContactsHelper = (get, applicationContext) => {
-  const partyType = get(state.form.partyType);
-  const user = applicationContext.getCurrentUser();
-
-  const { PARTY_TYPES } = applicationContext.getConstants();
-  const isExternalUser = applicationContext
-    .getUtilities()
-    .isExternalUser(user.role);
-
-  let showPaperPetitionEmailFieldAndConsentBox = !isExternalUser;
-
-  const contacts = getOptionsForContact({ PARTY_TYPES, partyType });
-
-  return { ...contacts, showPaperPetitionEmailFieldAndConsentBox };
-};
