+++ web-client/src/presenter/state.ts
@@ -12,7 +12,6 @@ import { applyStampFormHelper } from './computeds/applyStampFormHelper';
 import { batchDownloadHelper } from './computeds/batchDownloadHelper';
 import { blockedCasesReportHelper } from './computeds/blockedCasesReportHelper';
 import { caseDeadlineReportHelper } from './computeds/caseDeadlineReportHelper';
-import { caseDetailEditContactsHelper } from './computeds/caseDetailEditContactsHelper';
 import { caseDetailEditHelper } from './computeds/caseDetailEditHelper';
 import { caseDetailHeaderHelper } from './computeds/caseDetailHeaderHelper';
 import { caseDetailHelper } from './computeds/caseDetailHelper';
@@ -65,6 +64,7 @@ import { formattedWorkQueue } from './computeds/formattedWorkQueue';
 import { getConstants } from '../getConstants';
 import { getTrialCityName } from './computeds/formattedTrialCity';
 import { headerHelper } from './computeds/headerHelper';
+import { internalPetitionPartiesHelper } from './computeds/internalPetitionPartiesHelper';
 import { internalTypesHelper } from './computeds/internalTypesHelper';
 import { loadingHelper } from './computeds/loadingHelper';
 import { menuHelper } from './computeds/menuHelper';
@@ -98,7 +98,6 @@ import { showAppTimeoutModalHelper } from './computeds/showAppTimeoutModalHelper
 import { showSortableHeaders } from './computeds/showSortableHeaders';
 import { sortableColumnHelper } from './computeds/sortableColumnHelper';
 import { startCaseHelper } from './computeds/startCaseHelper';
-import { startCaseInternalContactsHelper } from './computeds/startCaseInternalContactsHelper';
 import { startCaseInternalHelper } from './computeds/startCaseInternalHelper';
 import { statisticsFormHelper } from './computeds/statisticsFormHelper';
 import { statisticsHelper } from './computeds/statisticsHelper';
@@ -133,7 +132,6 @@ const helpers = {
   batchDownloadHelper,
   blockedCasesReportHelper,
   caseDeadlineReportHelper,
-  caseDetailEditContactsHelper,
   caseDetailEditHelper,
   caseDetailHeaderHelper,
   caseDetailHelper,
@@ -183,6 +181,7 @@ const helpers = {
   formattedWorkQueue,
   getTrialCityName,
   headerHelper,
+  internalPetitionPartiesHelper,
   internalTypesHelper,
   loadingHelper,
   menuHelper,
@@ -216,7 +215,6 @@ const helpers = {
   showSortableHeaders,
   sortableColumnHelper,
   startCaseHelper,
-  startCaseInternalContactsHelper,
   startCaseInternalHelper,
   statisticsFormHelper,
   statisticsHelper,
